requirements.txt
langchain>=0.0.300
openai>=0.27.0
faiss-cpu>=1.7.4
chromadb>=0.3.30
pydantic>=1.10.7
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
matplotlib>=3.7.0
pytest>=7.4.0
uvicorn>=0.22.0
fastapi>=0.95.0
tenacity>=8.2.2
python-dateutil>=2.8.2
tqdm>=4.65.0
